backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
requests
python-dotenv
gunicorn